+++ CHANGELOG.md
@@ -5,6 +5,8 @@
         - order unsent reports by confirmed date
     - Bugfixes
 	- Application user in Docker container can't install packages. #2914
+    - UK:
+        - Added junction lookup, so you can search for things like "M60, Junction 2"
 
 * v3.0 (4th March 2020)
     - Security: